RDKCOM-3013: REFPLTV-1146 Bring up 64bit support in Rpi4 (for both kernel an...
Reason for change: In RPi4
Test Procedure: Build should be successful. rmfgeneric should build fine.
Risks: Low/None.
(cherry picked from commit e352da62137b1dfe079550fb94c91f248b289dfe)
(cherry picked from commit 1f2e8aed98784c50dd74fc6fbe5b867827af7698)
(cherry picked from commit 4b72b38eec82616e6259936e2a761a74ebc9427d)

+++ qamsource/simgr/oob/include/rmf_sicache.h
@@ -103,17 +103,17 @@ typedef enum rmf_SiServiceInformationType
 } rmf_SiServiceInformationType;
 
 /* PSI DB Supported handle types */
-typedef uint32_t rmf_SiTransportStreamHandle;
-typedef uint32_t rmf_SiProgramHandle;
-typedef uint32_t rmf_SiServiceComponentHandle;
-typedef uint32_t rmf_SiRatingDimensionHandle;
+typedef uintptr_t rmf_SiTransportStreamHandle;
+typedef uintptr_t rmf_SiProgramHandle;
+typedef uintptr_t rmf_SiServiceComponentHandle;
+typedef uintptr_t rmf_SiRatingDimensionHandle;
 
 /* SI DB Supported handle types */
-typedef uint32_t rmf_SiServiceHandle;
+typedef uintptr_t rmf_SiServiceHandle;
 typedef uint32_t rmf_SiServiceDetailsHandle;
-typedef uint32_t rmf_SiRatingDimensionHandle;
+typedef uintptr_t rmf_SiRatingDimensionHandle;
 
-typedef uint32_t rmf_SiGenericHandle; // Cast before use...
+typedef uintptr_t rmf_SiGenericHandle; // Cast before use...
 
 
 /**
